requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
rdkit
matplotlib
argparse
pillow
transformers
tqdm
argparse
tensorboard
jupyterlab
notebook
regex
xgboost
hypothesis
pytest
python-ta~=2.9.1
biopython
pytdc